templates/layers/python/azure-identity/fragments/requirements.txt
azure-identity>=1.19
